SITES-21278: Refactoring of code
